+++ examples/excelformer.py
@@ -1,13 +1,18 @@
-"""Reported (reproduced) accuracy(rmse for regression task) of ExcelFormer
+"""Reported (reproduced) accuracy (for multi-classification task), auc
+(for binary classification task) and rmse (for regression task)
 based on Table 1 of the paper https://arxiv.org/abs/2301.02819.
 ExcelFormer uses the same train-validation-test split as the Yandex paper.
+The reproduced results are based on Z-score Normalization, and the
+reported ones are based on :class:`QuantileTransformer` preprocessing
+in the Sklearn Python package. The above preprocessing is applied
+to numerical features.
 
-california_housing: 0.4587 (0.4733) num_layers=5, num_heads=4, num_layers=5,
-channels=32, lr: 0.001,
-jannis : 72.51 (72.38) num_heads=32, lr: 0.0001
-covtype: 97.17 (95.37)
-helena: 38.20 (36.80)
-higgs_small: 80.75 (65.17) lr: 0.0001
+california_housing: 0.4587 (0.4550) mixup: feature, num_layers: 3,
+gamma: 1.00, epochs: 300
+jannis : 72.51 (72.80) mixup: feature
+covtype: 97.17 (97.02) mixup: hidden
+helena: 38.20 (37.68) mixup: feature
+higgs_small: 80.75 (79.27) mixup: hidden
 """
 import argparse
 import os.path as osp
@@ -15,6 +20,7 @@ import os.path as osp
 import torch
 import torch.nn.functional as F
 from torch.optim.lr_scheduler import ExponentialLR
+from torchmetrics import AUROC, Accuracy, MeanSquaredError
 from tqdm import tqdm
 
 from torch_frame.data.loader import DataLoader
@@ -23,14 +29,16 @@ from torch_frame.nn import ExcelFormer
 from torch_frame.transforms import CatToNumTransform, MutualInformationSort
 
 parser = argparse.ArgumentParser()
-parser.add_argument('--dataset', type=str, default='higgs_small')
+parser.add_argument('--dataset', type=str, default='california_housing')
+parser.add_argument('--mixup', type=str, default=None,
+                    choices=[None, 'feature', 'hidden'])
 parser.add_argument('--channels', type=int, default=256)
 parser.add_argument('--batch_size', type=int, default=512)
 parser.add_argument('--num_heads', type=int, default=4)
 parser.add_argument('--num_layers', type=int, default=5)
 parser.add_argument('--lr', type=float, default=0.001)
+parser.add_argument('--gamma', type=float, default=0.95)
 parser.add_argument('--epochs', type=int, default=100)
-parser.add_argument('--mixup', type=bool, default=True)
 parser.add_argument('--compile', action='store_true')
 args = parser.parse_args()
 
@@ -76,6 +84,16 @@ if is_classification:
 else:
     out_channels = 1
 
+is_binary_class = is_classification and out_channels == 2
+
+if is_binary_class:
+    metric_computer = AUROC(task='binary')
+elif is_classification:
+    metric_computer = Accuracy(task='multiclass', num_classes=out_channels)
+else:
+    metric_computer = MeanSquaredError()
+metric_computer = metric_computer.to(device)
+
 model = ExcelFormer(
     in_channels=args.channels,
     out_channels=out_channels,
@@ -85,12 +103,13 @@ model = ExcelFormer(
     residual_dropout=0.,
     diam_dropout=0.3,
     aium_dropout=0.,
+    mixup=args.mixup,
     col_stats=mutual_info_sort.transformed_stats,
     col_names_dict=train_tensor_frame.col_names_dict,
 ).to(device)
 model = torch.compile(model, dynamic=True) if args.compile else model
 optimizer = torch.optim.Adam(model.parameters(), lr=args.lr)
-lr_scheduler = ExponentialLR(optimizer, gamma=0.95)
+lr_scheduler = ExponentialLR(optimizer, gamma=args.gamma)
 
 
 def train(epoch: int) -> float:
@@ -99,8 +118,10 @@ def train(epoch: int) -> float:
 
     for tf in tqdm(train_loader, desc=f'Epoch: {epoch}'):
         tf = tf.to(device)
-        pred_mixedup, y_mixedup = model.forward_mixup(tf)
+        # Train with FEAT-MIX or HIDDEN-MIX
+        pred_mixedup, y_mixedup = model(tf, mixup_encoded=True)
         if is_classification:
+            # Softly mixed one-hot labels
             loss = F.cross_entropy(pred_mixedup, y_mixedup)
         else:
             loss = F.mse_loss(pred_mixedup.view(-1), y_mixedup.view(-1))
@@ -115,29 +136,26 @@ def train(epoch: int) -> float:
 @torch.no_grad()
 def test(loader: DataLoader) -> float:
     model.eval()
-    accum = total_count = 0
-
+    metric_computer.reset()
     for tf in loader:
         tf = tf.to(device)
         pred = model(tf)
-        if is_classification:
+        if is_binary_class:
+            metric_computer.update(pred[:, 1], tf.y)
+        elif is_classification:
             pred_class = pred.argmax(dim=-1)
-            accum += float((tf.y == pred_class).sum())
+            metric_computer.update(pred_class, tf.y)
         else:
-            accum += float(
-                F.mse_loss(pred.view(-1), tf.y.view(-1), reduction='sum'))
-        total_count += len(tf.y)
+            metric_computer.update(pred.view(-1), tf.y.view(-1))
 
     if is_classification:
-        accuracy = accum / total_count
-        return accuracy
+        return metric_computer.compute().item()
     else:
-        rmse = (accum / total_count)**0.5
-        return rmse
+        return metric_computer.compute().item()**0.5
 
 
 if is_classification:
-    metric = 'Acc'
+    metric = 'Acc' if not is_binary_class else 'AUC'
     best_val_metric = 0
     best_test_metric = 0
 else: