+++ test/data/test_multi_nested_tensor.py
@@ -174,10 +174,25 @@ def test_multi_nested_tensor_basics(device):
     # Test row narrow
     assert_equal(tensor_mat[3:3 + 2],
                  multi_nested_tensor.narrow(dim=0, start=3, length=2))
+    assert len(multi_nested_tensor[0:0]) == 0
+    assert len(multi_nested_tensor[8:8]) == 0
 
     # Test multi_nested_tensor[List[int]] indexing
     for index in [[4], [2, 2], [-4, 1, 7], [3, -7, 1, 0], []]:
         multi_nested_tensor_indexed = multi_nested_tensor[index]
+        assert multi_nested_tensor_indexed.dtype == torch.long
+        assert multi_nested_tensor_indexed.shape[0] == len(index)
+        assert multi_nested_tensor_indexed.shape[1] == num_cols
+        for i, idx in enumerate(index):
+            for j in range(num_cols):
+                tensor = multi_nested_tensor_indexed[i, j]
+                assert torch.allclose(tensor_mat[idx][j], tensor)
+
+    # Test row-wise Boolean masking
+    for index in [[4], [2, 3], [0, 1, 7], []]:
+        mask = torch.zeros((num_rows, ), dtype=torch.bool, device=device)
+        mask[index] = True
+        multi_nested_tensor_indexed = multi_nested_tensor[mask]
         assert multi_nested_tensor_indexed.shape[0] == len(index)
         assert multi_nested_tensor_indexed.shape[1] == num_cols
         for i, idx in enumerate(index):
@@ -194,8 +209,17 @@ def test_multi_nested_tensor_basics(device):
 
     # Test column List[int] indexing
     for index in [[4], [2, 2], [-4, 1, 7], [3, -7, 1, 0], []]:
+        multi_nested_tensor_indexed = multi_nested_tensor[:, index]
         assert_equal(column_select(tensor_mat, index),
-                     multi_nested_tensor[:, index])
+                     multi_nested_tensor_indexed)
+        assert multi_nested_tensor_indexed.dtype == torch.long
+
+    # Test column-wise Boolean masking
+    for index in [[4], [2, 3], [0, 1, 7], []]:
+        mask = torch.zeros((num_cols, ), dtype=torch.bool, device=device)
+        mask[index] = True
+        assert_equal(column_select(tensor_mat, index),
+                     multi_nested_tensor[:, mask])
 
     # Test row range
     multi_nested_tensor_range = multi_nested_tensor[:, range(2, 6)]
@@ -203,6 +227,9 @@ def test_multi_nested_tensor_basics(device):
         for idx, j in enumerate(range(2, 6)):
             assert torch.allclose(tensor_mat[i][j],
                                   multi_nested_tensor_range[i, idx])
+
+    assert multi_nested_tensor[0:0, 1].shape == (0, 1, -1)
+    assert multi_nested_tensor[0:0, [1, 2]].shape == (0, 2, -1)
 
     # Test column slicing
     assert_equal(tensor_mat, multi_nested_tensor[:, :])
@@ -221,6 +248,7 @@ def test_multi_nested_tensor_basics(device):
     empty_multi_nested_tensor = multi_nested_tensor[:, 5:3]
     assert empty_multi_nested_tensor.shape[0] == num_rows
     assert empty_multi_nested_tensor.shape[1] == 0
+    assert empty_multi_nested_tensor.dtype == torch.long
 
     # Test column narrow
     assert_equal(column_select(tensor_mat, slice(3, 3 + 2)),